Can return dictionary that is inventory of store now. Customer is logged in, iterate over dictionary and add to order with amounts selected.

+++ StoreProject/StoreProject.Library/Product.cs
@@ -11,7 +11,7 @@ namespace StoreProject.Library
         public Product(string productName, decimal? price)
         {
             ProductName = productName;
-            Price = price;
+            Price = (decimal)price;
         }
 
         // Public properties to get or set the value of the products
